backend/requirements.txt
fastapi==0.103.1
peewee-aio==1.7.4
python-multipart==0.0.6
uvicorn==0.23.2
python-dotenv==1.0.0
aiomysql==0.1.0